Implemented basic functions for initializing skyline matrices.

+++ include/skyline.h
@@ -6,7 +6,7 @@
 
 skylineMatrix* initSkylineMatrix (int n);
 int addSkylineElement (skylineMatrix* matrix, int m, int n, double val);
-int getSkylineElement (skylineMatrix* matrix, int m, int n);
+double getSkylineElement (skylineMatrix* matrix, int m, int n);
 int updateSkylineElement (skylineMatrix* matrix, int m, int n, double val);
 
 #endif
